Refactor cli util into clowder_utils
