+++ clowder/clowder/cli/herd_controller.py
@@ -7,15 +7,15 @@
 
 from cement.ext.ext_argparse import ArgparseController, expose
 
-from clowder.clowder_controller import CLOWDER_CONTROLLER
-from clowder.cli.parallel import (
+from clowder.cli.parallel_commands import (
     herd,
     herd_parallel
 )
-from clowder.cli.util import options_help_message
+from clowder.clowder_controller import CLOWDER_CONTROLLER
 from clowder.clowder_repo import print_clowder_repo_status_fetch
-from clowder.util.decorators import valid_clowder_yaml_required
 from clowder.util.connectivity import network_connection_required
+from clowder.util.decorators import valid_clowder_yaml_required
+from clowder.util.clowder_utils import options_help_message
 
 
 class HerdController(ArgparseController):
