+++ clowder/clowder/cli/save_controller.py
@@ -15,12 +15,12 @@ from cement.ext.ext_argparse import ArgparseController, expose
 import clowder.util.formatting as fmt
 from clowder import ROOT_DIR
 from clowder.clowder_controller import CLOWDER_CONTROLLER
-from clowder.cli.util import (
+from clowder.clowder_repo import CLOWDER_REPO
+from clowder.util.decorators import valid_clowder_yaml_required
+from clowder.util.clowder_utils import (
     validate_groups,
     validate_projects_exist
 )
-from clowder.clowder_repo import CLOWDER_REPO
-from clowder.util.decorators import valid_clowder_yaml_required
 from clowder.yaml.saving import save_yaml
