+++ clowder/clowder/cli/start_controller.py
@@ -8,7 +8,10 @@
 from cement.ext.ext_argparse import ArgparseController, expose
 
 from clowder.clowder_controller import CLOWDER_CONTROLLER
-from clowder.cli.util import (
+from clowder.clowder_repo import print_clowder_repo_status
+from clowder.util.connectivity import network_connection_required
+from clowder.util.decorators import valid_clowder_yaml_required
+from clowder.util.clowder_utils import (
     filter_groups,
     filter_projects_on_project_names,
     options_help_message,
@@ -17,9 +20,6 @@ from clowder.cli.util import (
     validate_groups,
     validate_projects
 )
-from clowder.clowder_repo import print_clowder_repo_status
-from clowder.util.decorators import valid_clowder_yaml_required
-from clowder.util.connectivity import network_connection_required
 
 
 class StartController(ArgparseController):
