+++ clowder/clowder/util/clowder_utils.py
